requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.7.0
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-weasyprint==2.3.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
fonttools==4.53.1
html5lib==1.1
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
